requirements.txt
openai
nbformat
python-dotenv
